Big overhaul to some methods that take in Vector arguments
It's a bit of a mess.. but hopefully, they're relatively more usable.
These changes attempt to combat issues such as having a Vector with a
mixture of value-, reference-, or smart-pointer-types, resulting in a
vectors with type `Vector{Any}`. Declaring a catch-all that attempts to
convert its elements to `CxxRef`s in hopes they're all "ok" types is the
best I can do atm.