Change artifacts cache key to include julia version
Might not be the best idea when using the `nightly` string since nightly
might correspond to a slightly different version.  In theory, the
nightly version changes on a nightly basis, i.e. frequently.  For
correctness' sake, artifacts should probably not be cached for nightly.
In practice, however, it shouldn't be harmful.